Initial documentation of all classes
All classes which form part of the interface or which have objects
instantiated have been documented. The documentation is very sparse,
consisting largely of which methods exist, the barest skeleton of
what they do, and the types/numbers/names of their parameters. Also,
the innards have been reordered so that they are consistent both with
good practices (class methods first, then object methods) and with
each other (all methods are ordered the same).

There is a lot more that needs documented, not least of which is the
searching infrastructure. All classes have notes about what more needs
to be documented.